pyproject.toml
[tool.poetry]
name = "giftmanager"
version = "0.1.36"
description = ""
authors = ["Lester1989 <lester@techfak.uni-bielefeld.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
Jinja2 = "^3.1.3"
uvicorn = "^0.25.0"
SQLAlchemy = "^2.0.23"
alembic = "^1.13.1"
ulid-py = "^1.1.0"
fastapi-sqlalchemy = "^0.2.1"
pytest = "^7.4.3"
httpx = "^0.26.0"
pytailwindcss = "^0.2.0"
python-multipart = "^0.0.6"
psycopg2 = "^2.9.9"
resend = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
faker = "^22.6.0"
itsdangerous = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
